feat : 게시글 삭제 기능 인증/인가 구현

+++ src/main/java/com/study/proudcat/domain/post/service/PostService.java
@@ -148,9 +148,13 @@ public class PostService {
     }
 
     @Transactional
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId, String email) {
         log.info("PostService deletePost run..");
         Post post = getPostEntity(postId);
+
+        if (!post.isSameWriter(email)) {
+            throw new RestApiException(ErrorCode.NOT_PROPER_USER);
+        }
         postRepository.delete(post);
     }
